+++ custom_tools/partition_iterator.py
@@ -6,13 +6,12 @@ import random
 import json
 from typing import Dict, Tuple, Literal
 import time
-from datetime import timedelta
 
 import env_setup.global_config
 import config
 from env_setup import environment_setup
 from custom_tools import custom_arcpy
-from custom_tools.timing_decorator import timing_decorator
+from custom_tools.decorators.timing_decorator import timing_decorator
 
 from input_data import input_n50
 from file_manager.n100.file_manager_buildings import Building_N100
