+++ generalization/n100/building/calculate_point_values.py
@@ -1,7 +1,6 @@
 # Importing custom files relative to the root path
 from custom_tools import custom_arcpy
 from env_setup import environment_setup
-from input_data import input_n50
 from file_manager.n100.file_manager_buildings import Building_N100
 from constants.n100_constants import N100_SQLResources
 
@@ -10,7 +9,7 @@ import arcpy
 from collections import Counter
 
 # Importing timing decorator
-from custom_tools.timing_decorator import timing_decorator
+from custom_tools.decorators.timing_decorator import timing_decorator
 
 
 @timing_decorator
