+++ generalization/n100/building/polygon_resolve_building_conflicts.py
@@ -2,7 +2,6 @@
 import arcpy
 
 # Importing custom modules
-import config
 import input_data.input_n50
 import input_data.input_n100
 from custom_tools import custom_arcpy
@@ -17,9 +16,8 @@ from env_setup import environment_setup
 from file_manager.n100.file_manager_buildings import Building_N100
 
 # Importing timing decorator
-from custom_tools.timing_decorator import timing_decorator
 
-from custom_tools.timing_decorator import timing_decorator
+from custom_tools.decorators.timing_decorator import timing_decorator
 
 
 @timing_decorator
