+++ generalization/n100/building/polygon_to_point.py
@@ -1,9 +1,7 @@
 # Importing modules
 import arcpy
-import time
 
 # Importing custom files
-from custom_tools import custom_arcpy
 from env_setup import environment_setup
 
 # Importing temporary files
@@ -11,7 +9,7 @@ from file_manager.n100.file_manager_buildings import Building_N100
 
 
 # Importing timing decorator
-from custom_tools.timing_decorator import timing_decorator
+from custom_tools.decorators.timing_decorator import timing_decorator
 
 
 @timing_decorator
