+++ generalization/n100/building/removing_overlapping_points.py
@@ -1,16 +1,13 @@
 # Importing modules
-import arcpy
 
 # Importing custom files
-from file_manager.n100.file_manager_buildings import Building_N100
 
 # Import custom modules
 from custom_tools import custom_arcpy
 from env_setup import environment_setup
-from custom_tools.timing_decorator import timing_decorator
 from custom_tools.compare_feature_classes import compare_feature_classes
 from custom_tools.polygon_processor import PolygonProcessor
-from constants.n100_constants import N100_Symbology, N100_SQLResources, N100_Values
+from constants.n100_constants import N100_Symbology
 
 # Importing custom files
 from file_manager.n100.file_manager_buildings import Building_N100
@@ -20,7 +17,7 @@ from constants.n100_constants import N100_SQLResources
 import arcpy
 
 # Importing timing decorator
-from custom_tools.timing_decorator import timing_decorator
+from custom_tools.decorators.timing_decorator import timing_decorator
 
 
 @timing_decorator
