+++ generalization/n100/building/removing_points_and_erasing_polygons_in_water_features.py
@@ -9,7 +9,7 @@ from input_data.input_symbology import SymbologyN100
 # Import custom modules
 from custom_tools import custom_arcpy
 from env_setup import environment_setup
-from custom_tools.timing_decorator import timing_decorator
+from custom_tools.decorators.timing_decorator import timing_decorator
 
 
 @timing_decorator
